app/requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
boto3==1.14.12
botocore==1.17.12
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Django==3.0.3
django-environ==0.4.5
django-markdownx==3.0.1
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
google-api-python-client==1.7.12
google-auth==1.11.2
google-auth-httplib2==0.0.3
gunicorn==20.0.4
httplib2==0.17.0
idna==2.9
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
Pillow==7.1.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
rope==0.16.0
rsa==4.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
wrapt==1.11.2
sentry-sdk==0.16.1